requirements.txt
appdirs==1.4.4
APScheduler==3.10.4
backtrader==1.9.78.123
bcrypt==4.1.2
beautifulsoup4==4.13.4
black==24.2.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.0
cryptography==43.0.1
curl_cffi==0.11.1
docker==7.0.0
ecdsa==0.19.1
flake8==7.0.0
Flask==3.0.2
Flask-Cors==4.0.0
Flask-JWT-Extended==4.7.1
Flask-SQLAlchemy==3.1.1
frozendict==2.4.6
google-auth==2.40.1
google-auth-oauthlib==1.2.2
gunicorn==21.2.0
html5lib==1.1
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==5.4.0
MarkupSafe==3.0.2
mccabe==0.7.0
multitasking==0.0.11
mypy_extensions==1.1.0
numpy
oauthlib==3.2.2
packaging==25.0
pandas
pathspec==0.12.1
peewee==3.18.1
platformdirs==4.3.8
pluggy==1.6.0
protobuf==6.31.0
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.11.1
pycparser==2.22
pyflakes==3.2.0
PyJWT==2.10.1
pytest==8.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
pytz==2025.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
soupsieve==2.7
SQLAlchemy==2.0.28
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.4.0
webencodings==0.5.1
websockets==15.0.1
Werkzeug==3.1.3
pytest-flask==1.3.0
pytest-mock==3.12.0
pytest-cov==4.0.0
factory-boy==3.3.0
openai>=1.0.0
pandas-ta
ccxt>=4.0.0
eth-account>=0.8.0
hyperliquid-python-sdk>=0.21.0
